conclusão das permissões

+++ loja/resources/views/users/index.blade.php
@@ -8,7 +8,9 @@
             <h2>Usuários</h2>
         </div>
         <div class="pull-right">
+            @can('user-create')
             <a class="btn btn-success" href="{{ route('users.create') }}"> + Novo usuário</a>
+            @endcan
         </div>
     </div>
 </div>
@@ -56,13 +58,16 @@
 
     <td>
        <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Mostrar</a>
+       @can('user-edit')
        <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Editar</a>
-
+       @endcan
+       @can('user-delete')
         {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
 
             {!! Form::submit('Apagar', ['class' => 'btn btn-danger']) !!}
 
         {!! Form::close() !!}
+        @endcan
 
     </td>
   </tr>
